requirements.txt
tqdm
gdown
json-tricks
torch
opencv-python
prettytable
scipy
torchvision
numpy==1.26.4
matplotlib
easydict
PyYAML
tensorboardX
vedo
imageio
huggingface_hub
ultralytics==8.1.46
timm==0.8.13.dev0
lapx>=0.5.2
confluent_kafka
pillow
gdown
tensorrt==10.1.0
fastapi
uvicorn
python-multipart
opencv-contrib-python-headless
PyQt5
requests
pyrealsense2
